# api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8183/elastic-query-service
  description: Generated server url
paths:
  /documents/get-document-by-text:
    post:
      tags:
      - elastic-document-controller
      summary: Get elastic document by text.
      operationId: getDocumentByText
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ElasticQueryServiceRequestModel'
        required: true
      responses:
        "200":
          description: Successful response.
          content:
            application/vnd.api.v1+json:
              schema:
                $ref: '#/components/schemas/ElasticQueryServiceResponseModel'
        "500":
          description: Internal server error.
          content:
            application/vnd.api.v1+json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ElasticQueryServiceResponseModel'
        "400":
          description: Not found.
          content:
            application/vnd.api.v1+json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ElasticQueryServiceResponseModel'
  /documents:
    get:
      tags:
      - elastic-document-controller
      summary: Get all elastic documents!
      operationId: getAllDocuments
      responses:
        "200":
          description: Success
          content:
            application/vnd.api.v1+json:
              schema:
                $ref: '#/components/schemas/ElasticQueryServiceResponseModel'
        "500":
          description: Internal server error.
          content:
            application/vnd.api.v1+json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ElasticQueryServiceResponseModel'
        "400":
          description: Not found.
          content:
            application/vnd.api.v1+json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ElasticQueryServiceResponseModel'
  /documents/{id}:
    get:
      tags:
      - elastic-document-controller
      summary: Get elastic document by id.
      operationId: getDocumentByIdV2_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful response.
          content:
            application/vnd.api.v2+json:
              schema:
                $ref: '#/components/schemas/ElasticQueryServiceResponseModelV2'
            application/vnd.api.v1+json:
              schema:
                $ref: '#/components/schemas/ElasticQueryServiceResponseModel'
        "500":
          description: Internal server error.
          content:
            application/vnd.api.v2+json:
              schema:
                $ref: '#/components/schemas/ElasticQueryServiceResponseModelV2'
        "400":
          description: Not found.
          content:
            application/vnd.api.v2+json:
              schema:
                $ref: '#/components/schemas/ElasticQueryServiceResponseModelV2'
components:
  schemas:
    ElasticQueryServiceRequestModel:
      required:
      - text
      type: object
      properties:
        id:
          type: string
        text:
          type: string
    ElasticQueryServiceResponseModel:
      type: object
      properties:
        id:
          type: string
        userId:
          type: integer
          format: int64
        text:
          type: string
        createdAt:
          type: string
          format: date-time
        links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
    Link:
      type: object
      properties:
        rel:
          type: string
        href:
          type: string
        hreflang:
          type: string
        media:
          type: string
        title:
          type: string
        type:
          type: string
        deprecation:
          type: string
        profile:
          type: string
        name:
          type: string
    ElasticQueryServiceResponseModelV2:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        text:
          type: string
        text2:
          type: string
        links:
          type: array
          items:
            $ref: '#/components/schemas/Link'